Fix variable extraction when string uses variables concatenation
