+++ template/template.go
@@ -271,102 +271,6 @@ func Substitute(template string, mapping Mapping) (string, error) {
 	return SubstituteWith(template, mapping, DefaultPattern)
 }
 
-// ExtractVariables returns a map of all the variables defined in the specified
-// composefile (dict representation) and their default value if any.
-func ExtractVariables(configDict map[string]interface{}, pattern *regexp.Regexp) map[string]Variable {
-	if pattern == nil {
-		pattern = DefaultPattern
-	}
-	return recurseExtract(configDict, pattern)
-}
-
-func recurseExtract(value interface{}, pattern *regexp.Regexp) map[string]Variable {
-	m := map[string]Variable{}
-
-	switch value := value.(type) {
-	case string:
-		if values, is := extractVariable(value, pattern); is {
-			for _, v := range values {
-				m[v.Name] = v
-			}
-		}
-	case map[string]interface{}:
-		for _, elem := range value {
-			submap := recurseExtract(elem, pattern)
-			for key, value := range submap {
-				m[key] = value
-			}
-		}
-
-	case []interface{}:
-		for _, elem := range value {
-			if values, is := extractVariable(elem, pattern); is {
-				for _, v := range values {
-					m[v.Name] = v
-				}
-			}
-		}
-	}
-
-	return m
-}
-
-type Variable struct {
-	Name          string
-	DefaultValue  string
-	PresenceValue string
-	Required      bool
-}
-
-func extractVariable(value interface{}, pattern *regexp.Regexp) ([]Variable, bool) {
-	sValue, ok := value.(string)
-	if !ok {
-		return []Variable{}, false
-	}
-	matches := pattern.FindAllStringSubmatch(sValue, -1)
-	if len(matches) == 0 {
-		return []Variable{}, false
-	}
-	values := []Variable{}
-	for _, match := range matches {
-		groups := matchGroups(match, pattern)
-		if escaped := groups[groupEscaped]; escaped != "" {
-			continue
-		}
-		val := groups[groupNamed]
-		if val == "" {
-			val = groups[groupBraced]
-		}
-		name := val
-		var defaultValue string
-		var presenceValue string
-		var required bool
-		switch {
-		case strings.Contains(val, ":?"):
-			name, _ = partition(val, ":?")
-			required = true
-		case strings.Contains(val, "?"):
-			name, _ = partition(val, "?")
-			required = true
-		case strings.Contains(val, ":-"):
-			name, defaultValue = partition(val, ":-")
-		case strings.Contains(val, "-"):
-			name, defaultValue = partition(val, "-")
-		case strings.Contains(val, ":+"):
-			name, presenceValue = partition(val, ":+")
-		case strings.Contains(val, "+"):
-			name, presenceValue = partition(val, "+")
-		}
-		values = append(values, Variable{
-			Name:          name,
-			DefaultValue:  defaultValue,
-			PresenceValue: presenceValue,
-			Required:      required,
-		})
-	}
-	return values, len(values) > 0
-}
-
 // Soft default (fall back if unset or empty)
 func defaultWhenEmptyOrUnset(substitution string, mapping Mapping) (string, bool, error) {
 	return withDefaultWhenAbsence(substitution, mapping, true)
