+++ template/template_test.go
@@ -405,6 +405,10 @@ func TestSubstituteWithReplacementFunc(t *testing.T) {
 
 	_, err = SubstituteWithOptions("ok ${NOTHERE}", defaultMapping, options...)
 	assert.Check(t, is.ErrorContains(err, "bad choice"))
+
+	result, err = SubstituteWith("ok ${SUBDOMAIN:-redis}.${FOO:?}", defaultMapping, DefaultPattern)
+	assert.NilError(t, err)
+	assert.Check(t, is.Equal("ok redis.first", result))
 }
 
 func TestSubstituteWithReplacementAppliedFunc(t *testing.T) {
@@ -475,128 +479,6 @@ func TestPrecedence(t *testing.T) {
 		result, err := Substitute(tc.template, defaultMapping)
 		assert.Check(t, is.DeepEqual(tc.err, err))
 		assert.Check(t, is.Equal(tc.expected, result))
-	}
-}
-
-func TestExtractVariables(t *testing.T) {
-	testCases := []struct {
-		name     string
-		dict     map[string]interface{}
-		expected map[string]Variable
-	}{
-		{
-			name:     "empty",
-			dict:     map[string]interface{}{},
-			expected: map[string]Variable{},
-		},
-		{
-			name: "no-variables",
-			dict: map[string]interface{}{
-				"foo": "bar",
-			},
-			expected: map[string]Variable{},
-		},
-		{
-			name: "variable-without-curly-braces",
-			dict: map[string]interface{}{
-				"foo": "$bar",
-			},
-			expected: map[string]Variable{
-				"bar": {Name: "bar"},
-			},
-		},
-		{
-			name: "variable",
-			dict: map[string]interface{}{
-				"foo": "${bar}",
-			},
-			expected: map[string]Variable{
-				"bar": {Name: "bar", DefaultValue: ""},
-			},
-		},
-		{
-			name: "required-variable",
-			dict: map[string]interface{}{
-				"foo": "${bar?:foo}",
-			},
-			expected: map[string]Variable{
-				"bar": {Name: "bar", DefaultValue: "", Required: true},
-			},
-		},
-		{
-			name: "required-variable2",
-			dict: map[string]interface{}{
-				"foo": "${bar?foo}",
-			},
-			expected: map[string]Variable{
-				"bar": {Name: "bar", DefaultValue: "", Required: true},
-			},
-		},
-		{
-			name: "default-variable",
-			dict: map[string]interface{}{
-				"foo": "${bar:-foo}",
-			},
-			expected: map[string]Variable{
-				"bar": {Name: "bar", DefaultValue: "foo"},
-			},
-		},
-		{
-			name: "default-variable2",
-			dict: map[string]interface{}{
-				"foo": "${bar-foo}",
-			},
-			expected: map[string]Variable{
-				"bar": {Name: "bar", DefaultValue: "foo"},
-			},
-		},
-		{
-			name: "multiple-values",
-			dict: map[string]interface{}{
-				"foo": "${bar:-foo}",
-				"bar": map[string]interface{}{
-					"foo": "${fruit:-banana}",
-					"bar": "vegetable",
-				},
-				"baz": []interface{}{
-					"foo",
-					"$docker:${project:-cli}",
-					"$toto",
-				},
-			},
-			expected: map[string]Variable{
-				"bar":     {Name: "bar", DefaultValue: "foo"},
-				"fruit":   {Name: "fruit", DefaultValue: "banana"},
-				"toto":    {Name: "toto", DefaultValue: ""},
-				"docker":  {Name: "docker", DefaultValue: ""},
-				"project": {Name: "project", DefaultValue: "cli"},
-			},
-		},
-		{
-			name: "presence-value-nonEmpty",
-			dict: map[string]interface{}{
-				"foo": "${bar:+foo}",
-			},
-			expected: map[string]Variable{
-				"bar": {Name: "bar", PresenceValue: "foo"},
-			},
-		},
-		{
-			name: "presence-value",
-			dict: map[string]interface{}{
-				"foo": "${bar+foo}",
-			},
-			expected: map[string]Variable{
-				"bar": {Name: "bar", PresenceValue: "foo"},
-			},
-		},
-	}
-	for _, tc := range testCases {
-		tc := tc
-		t.Run(tc.name, func(t *testing.T) {
-			actual := ExtractVariables(tc.dict, DefaultPattern)
-			assert.Check(t, is.DeepEqual(actual, tc.expected))
-		})
 	}
 }
